Integration tests are being implemented
