+++ GodelTech.Microservices.Security.sln
@@ -11,6 +11,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GodelTech.Microservices.Api
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GodelTech.Microservices.Security.IntegrationTests", "test\GodelTech.Microservices.Security.IntegrationTests\GodelTech.Microservices.Security.IntegrationTests.csproj", "{7537770D-9C9D-4D9D-86FF-57BD12C0D17B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GodeTech.Microservice.IdentityProviderWebsite", "test\GodeTech.Microservice.IdentityProviderWebsite\GodeTech.Microservice.IdentityProviderWebsite.csproj", "{E0FFDC9B-E6F5-4307-94F1-7618CDA172EA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{7537770D-9C9D-4D9D-86FF-57BD12C0D17B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7537770D-9C9D-4D9D-86FF-57BD12C0D17B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7537770D-9C9D-4D9D-86FF-57BD12C0D17B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E0FFDC9B-E6F5-4307-94F1-7618CDA172EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E0FFDC9B-E6F5-4307-94F1-7618CDA172EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E0FFDC9B-E6F5-4307-94F1-7618CDA172EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E0FFDC9B-E6F5-4307-94F1-7618CDA172EA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
