+++ src/GodelTech.Microservices.Security/GodelTech.Microservices.Security.csproj
@@ -17,9 +17,9 @@
 
   <ItemGroup>
 
-    <PackageReference Include="GodelTech.Microservices.Core" Version="1.1.0" />
+    <PackageReference Include="GodelTech.Microservices.Core" Version="1.1.1" />
 
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.6" />
 
   </ItemGroup>
