+++ test/GodeTech.Microservice.IdentityProviderWebsite/GodeTech.Microservice.IdentityProviderWebsite.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+</Project>
