+++ test/GodeTech.Microservice.IdentityProviderWebsite/Program.cs
@@ -1,7 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
-namespace GodelTech.Microservices.ApiWebsite
+namespace GodeTech.Microservice.IdentityProviderWebsite
 {
     public class Program
     {
