+++ test/GodeTech.Microservice.IdentityProviderWebsite/appsettings.json
@@ -6,12 +6,5 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-
-  "ApiSecurityConfig": {
-    "AuthorityUri": "http://localhost:7000",
-    "Issuer": "http://godeltech",
-    "Audience": "apiwebsite"
-  },
-
   "AllowedHosts": "*"
 }
