+++ test/GodelTech.Microservices.ApiWebsite/GodelTech.Microservices.ApiWebsite.csproj
@@ -4,5 +4,9 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\GodelTech.Microservices.Security\GodelTech.Microservices.Security.csproj" />
+  </ItemGroup>
+
 
 </Project>
